+++ toys-core/examples/multi-file-gem/.toys.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+xpand :clean, paths: ["pkg"]
+
+expand :gem_build
+expand :gem_build, name: "release", push_gem: true
+expand :gem_build, name: "install", install_gem: true