+++ toys-core/examples/multi-file-gem/tools/greet.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+desc "Display a simple greeting"
+flag :whom, default: "world"
+def run
+  puts "Hello, #{whom}!"
+end