+++ toys/CHANGELOG.md
@@ -5,6 +5,7 @@
 * ADDED: Tab completion for bash.
 * ADDED: Tools can provide an interrupt handler.
 * ADDED: Flag handlers can accept the symbolic names `:set` and `:push`.
+* ADDED: `:gem_build` template includes an `:install_gem` options.
 * IMPROVED: Toys no longer displays a stack trace if a tool is interrupted.
 * FIXED: Prevented toys-core from being ousted from the load path if a toys file invoked bundler setup.